requirements.txt
streamlit==1.30.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
streamlit-autorefresh==1.0.1
plotly==5.17.0